# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for BAST Request System.",
        "title": "BAST Request API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/audit-logs": {
            "get": {
                "description": "Retrieve a list of audit logs, optionally filtered by parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit-logs"
                ],
                "summary": "Get all audit logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Table Name filter",
                        "name": "table_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Record ID filter",
                        "name": "record_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Performed By filter",
                        "name": "performed_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date From filter (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date To filter (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/bast-request_internal_models.AuditLog"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/audit-logs/{id}": {
            "get": {
                "description": "Retrieve a single audit log based on its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit-logs"
                ],
                "summary": "Get an audit log by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Audit Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bast-request_internal_models.AuditLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Login user to get JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bast-request_internal_models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register a new user with username, email, password, and role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Register Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bast-request_internal_models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/bast-formats": {
            "get": {
                "description": "Retrieve a list of all BAST formats",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bast-formats"
                ],
                "summary": "Get all BAST formats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/bast-request_internal_models.BastFormat"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new BAST format to the master data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bast-formats"
                ],
                "summary": "Create a new BAST format",
                "parameters": [
                    {
                        "description": "Format Data",
                        "name": "format",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bast-request_internal_models.BastFormat"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/bast-request_internal_models.BastFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/bast-formats/{id}": {
            "get": {
                "description": "Retrieve a single BAST format based on its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bast-formats"
                ],
                "summary": "Get a BAST format by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Format ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bast-request_internal_models.BastFormat"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update data for a specific BAST format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bast-formats"
                ],
                "summary": "Update an existing BAST format",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Format ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Format Data",
                        "name": "format",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bast-request_internal_models.BastFormat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bast-request_internal_models.BastFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Nonaktifkan BAST format based on ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bast-formats"
                ],
                "summary": "Delete a BAST format",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Format ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "bast-request_internal_models.AuditLog": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "POST, PUT, DELETE, PATCH",
                    "type": "string"
                },
                "auditLogID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "newData": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "oldData": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "performedBy": {
                    "type": "string"
                },
                "recordID": {
                    "type": "string"
                },
                "targetTable": {
                    "type": "string"
                }
            }
        },
        "bast-request_internal_models.BastFormat": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "formatID": {
                    "type": "string"
                },
                "formatName": {
                    "type": "string"
                },
                "formatPattern": {
                    "description": "e.g. BAST/{YYYY}/{MM}/{SEQ}",
                    "type": "string"
                },
                "formatType": {
                    "description": "PO, Internal",
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "bast-request_internal_models.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "bast-request_internal_models.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "role",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}